// src/services/wallets/interfaces/vault.ts
export interface VaultPresentationData extends WalletPresentationData {}

export interface VaultSpecs {
  xpubs: string[]; // signers' xpubs
  nextFreeAddressIndex: number; // external-chain free address marker
  nextFreeChangeAddressIndex: number; // internal-chain free address marker
  totalExternalAddresses: number; // total number of external addresses the user generated (starts at 1)
  receivingAddress?: string; // current receiving address(external chain)
  addresses?: AddressCache; // cached addresses
  addressPubs?: AddressPubs; // cached pubs
  confirmedUTXOs: UTXO[]; // utxo set available for use
  unconfirmedUTXOs: UTXO[]; // utxos to arrive
  balances: Balances; // confirmed/unconfirmed balances
  transactions: Transaction[]; // transactions belonging to this wallet
  hasNewUpdates: boolean; // spec vars have a new update?
  lastSynched: number; // vault's last sync timestamp
}
// ...
export interface MiniscriptScheme {
  miniscriptElements: MiniscriptElements;
  keyInfoMap: KeyInfoMap;
  miniscriptPolicy: string; // miniscript policy
  miniscript: string; // miniscript
  usedMiniscriptTypes: MiniscriptTypes[];
}

export interface VaultScheme {
  m: number; // threshold number of signatures required
  n: number; // total number of xpubs
  multisigScriptType?: MultisigScriptType; // multisig script type(allows for more complex and flexible vaults)
  miniscriptScheme?: MiniscriptScheme;
}
// ...
export type RegisteredVaultInfo = {
  vaultId: string;
  registered: boolean;
  registrationInfo?: string;
  hmac?: string;
};

export interface VaultSigner {
  // Represents xpub(Extended Key) belonging to one of the Signers,
  // Rel: VaultSigner(Extended Key) could only belong to one Signer, and is an active part of a Vault(s)
  masterFingerprint: string;
  xpub: string;
  xpriv?: string;
  xfp: string;
  derivationPath: string;
  registeredVaults?: RegisteredVaultInfo[];
}

export interface Vault {
  // Represents a Vault
  // Rel: Created using multiple VaultSigners(Extended Keys)
  id: string; // vault identifier(derived from xpub)
  entityKind: EntityKind; // Vault vs Wallet identifier
  type: VaultType; // type of vault
  networkType: NetworkType; // testnet/mainnet
  isMultiSig: boolean; // true
  scheme: VaultScheme; // scheme of vault(m of n)
  signers: VaultSigner[]; // signers of the vault
  presentationData: VaultPresentationData;
  specs: VaultSpecs;
  archived: boolean;
  isMigrating?: boolean;
  archivedId?: string;
  scriptType: ScriptTypes;
  receivingAddress?: string;
}